// src/models/OrderStatusModel.ts
enum OrderStatusModel {
  NEW = 'NEW',
  // CANCELED = 'CANCELED',
  PROCESSING = 'PROCESSING',
  PACKAGED = 'PACKAGED',
  PICKED = 'PICKED',
  SHIPPING = 'SHIPPING',
  DELIVERED = 'DELIVERED',
  // RETURN_REQUESTED = 'RETURN_REQUESTED',
  // RETURNED = 'RETURNED',
  PAID = 'PAID',
  // REFUNDED = 'REFUNDED',
}

export default OrderStatusModel;